move gitlab to a separate package too
