+++ .changeset/pink-chicken-smile.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-backend-module-github': minor
+---
+
+Added package, moving out GitHub specific functionality from the catalog-backend
