+++ plugins/catalog-backend-module-github/README.md
@@ -1,0 +1,8 @@
+# Catalog Backend Module for GitHub
+
+This is an extension module to the plugin-catalog-backend plugin, providing extensions targeted at GitHub offerings.
+
+## Getting started
+
+See [Backstage documentation](https://backstage.io/docs/integrations/github/discovery) for details on how to install
+and configure the plugin.
