+++ plugins/catalog-backend-module-github/api-report.md
@@ -1,0 +1,126 @@
+## API Report File for "@backstage/plugin-catalog-backend-module-github"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { CatalogProcessor } from '@backstage/plugin-catalog-backend';
+import { CatalogProcessorEmit } from '@backstage/plugin-catalog-backend';
+import { Config } from '@backstage/config';
+import { EntityProvider } from '@backstage/plugin-catalog-backend';
+import { EntityProviderConnection } from '@backstage/plugin-catalog-backend';
+import { GithubCredentialsProvider } from '@backstage/integration';
+import { GitHubIntegrationConfig } from '@backstage/integration';
+import { LocationSpec } from '@backstage/plugin-catalog-backend';
+import { Logger } from 'winston';
+import { ScmIntegrationRegistry } from '@backstage/integration';
+
+// @public
+export class GithubDiscoveryProcessor implements CatalogProcessor {
+  constructor(options: {
+    integrations: ScmIntegrationRegistry;
+    logger: Logger;
+    githubCredentialsProvider?: GithubCredentialsProvider;
+  });
+  // (undocumented)
+  static fromConfig(
+    config: Config,
+    options: {
+      logger: Logger;
+      githubCredentialsProvider?: GithubCredentialsProvider;
+    },
+  ): GithubDiscoveryProcessor;
+  // (undocumented)
+  getProcessorName(): string;
+  // (undocumented)
+  readLocation(
+    location: LocationSpec,
+    _optional: boolean,
+    emit: CatalogProcessorEmit,
+  ): Promise<boolean>;
+}
+
+// @public
+export type GithubMultiOrgConfig = Array<{
+  name: string;
+  groupNamespace: string;
+  userNamespace: string | undefined;
+}>;
+
+// @public
+export class GithubMultiOrgReaderProcessor implements CatalogProcessor {
+  constructor(options: {
+    integrations: ScmIntegrationRegistry;
+    logger: Logger;
+    orgs: GithubMultiOrgConfig;
+    githubCredentialsProvider?: GithubCredentialsProvider;
+  });
+  // (undocumented)
+  static fromConfig(
+    config: Config,
+    options: {
+      logger: Logger;
+      githubCredentialsProvider?: GithubCredentialsProvider;
+    },
+  ): GithubMultiOrgReaderProcessor;
+  // (undocumented)
+  getProcessorName(): string;
+  // (undocumented)
+  readLocation(
+    location: LocationSpec,
+    _optional: boolean,
+    emit: CatalogProcessorEmit,
+  ): Promise<boolean>;
+}
+
+// @public (undocumented)
+export class GitHubOrgEntityProvider implements EntityProvider {
+  constructor(options: {
+    id: string;
+    orgUrl: string;
+    gitHubConfig: GitHubIntegrationConfig;
+    logger: Logger;
+    githubCredentialsProvider?: GithubCredentialsProvider;
+  });
+  // (undocumented)
+  connect(connection: EntityProviderConnection): Promise<void>;
+  // (undocumented)
+  static fromConfig(
+    config: Config,
+    options: {
+      id: string;
+      orgUrl: string;
+      logger: Logger;
+      githubCredentialsProvider?: GithubCredentialsProvider;
+    },
+  ): GitHubOrgEntityProvider;
+  // (undocumented)
+  getProviderName(): string;
+  // (undocumented)
+  read(): Promise<void>;
+}
+
+// @public
+export class GithubOrgReaderProcessor implements CatalogProcessor {
+  constructor(options: {
+    integrations: ScmIntegrationRegistry;
+    logger: Logger;
+    githubCredentialsProvider?: GithubCredentialsProvider;
+  });
+  // (undocumented)
+  static fromConfig(
+    config: Config,
+    options: {
+      logger: Logger;
+      githubCredentialsProvider?: GithubCredentialsProvider;
+    },
+  ): GithubOrgReaderProcessor;
+  // (undocumented)
+  getProcessorName(): string;
+  // (undocumented)
+  readLocation(
+    location: LocationSpec,
+    _optional: boolean,
+    emit: CatalogProcessorEmit,
+  ): Promise<boolean>;
+}
+```
