+++ plugins/catalog-backend-module-github/src/lib/config.ts
