+++ plugins/catalog-backend/package.json
@@ -46,7 +46,6 @@
     "@backstage/plugin-scaffolder-common": "^0.3.0-next.0",
     "@backstage/plugin-search-common": "^0.3.1-next.0",
     "@backstage/types": "^0.1.3",
-    "@octokit/graphql": "^4.5.8",
     "@types/express": "^4.17.6",
     "codeowners-utils": "^1.0.2",
     "core-js": "^3.6.5",
